// src/SupplierPendingProducts/Types/SupplierPendingProductTypes.tsx
export interface SupplierPendingProductTypes {
  id: number;
  supplierId: number;
  supplierName: string;
  warehouseId: number;
  warehouseName: string;
  productId: number;
  productName: string;
  purchasePrice: number;
  purchaseStatus: string;
  remainingAmount: number;
  quantity: number;
  observation: string,
  date: string;
  total:number;
  status: string;
}